#2 初期化をロジックに切り出し
